% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_redundancy.R
\name{calculate_redundancy}
\alias{calculate_redundancy}
\title{calculate_redundancy}
\usage{
calculate_redundancy(annotation_table)
}
\arguments{
\item{annotation_table}{annotation table}
}
\value{
redundancy
}
\description{
calculate_redundancy
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
