% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-data_checking.R
\name{check4fmsea}
\alias{check4fmsea}
\title{Check data format for fMSEA}
\usage{
check4fmsea(object_list)
}
\arguments{
\item{object_list}{object list, the items should be mass_dataset class from massdataset}
}
\value{
Checking results
}
\description{
Check data format for fMSEA
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
