% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_conflicts.R
\name{fmsea_conflicts}
\alias{fmsea_conflicts}
\title{Conflicts between the fmsea and other packages}
\usage{
fmsea_conflicts()
}
\value{
fmsea_conflicts
}
\description{
This function lists all the conflicts between packages in the fmsea
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
fmsea_conflicts()
}
