% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_fmsea_logo.R
\name{fmsea_logo}
\alias{fmsea_logo}
\title{Show the base information of fmsea pacakge}
\usage{
fmsea_logo()
}
\value{
A ASCII log of fmsea
}
\description{
Show the base information of fmsea pacakge.
}
\examples{
fmsea_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
