% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_utils.R
\name{fmsea_packages}
\alias{fmsea_packages}
\title{List all packages in the fmsea}
\usage{
fmsea_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include fmsea in the list?}
}
\value{
fmsea_packages
}
\description{
List all packages in the fmsea
}
\examples{
fmsea_packages()
}
